+++ src/com/intellij/idea/plugin/hybris/toolwindow/loggers/tree/LoggersOptionsTree.kt
@@ -28,7 +28,6 @@ import com.intellij.ui.tree.AsyncTreeModel
 import com.intellij.ui.treeStructure.Tree
 import com.intellij.util.containers.Convertor
 import java.io.Serial
-import javax.swing.event.TreeModelListener
 import javax.swing.tree.DefaultMutableTreeNode
 import javax.swing.tree.TreePath
 
@@ -55,7 +54,6 @@ class LoggersOptionsTree(val myProject: Project) : Tree(), Disposable {
 
     fun update(connections: Map<RemoteConnectionSettings, Boolean>) = myTreeModel.reload(connections)
     fun update() = myTreeModel.reload()
-    fun addTreeModelListener(listener: TreeModelListener) = model.addTreeModelListener(listener)
 
     companion object {
         @Serial