+++ src/com/intellij/idea/plugin/hybris/toolwindow/system/bean/components/BSTreePanel.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,9 +23,12 @@ import com.intellij.idea.plugin.hybris.system.bean.meta.BSGlobalMetaModel
 import com.intellij.idea.plugin.hybris.toolwindow.system.bean.forms.BSMetaBeanView
 import com.intellij.idea.plugin.hybris.toolwindow.system.bean.forms.BSMetaEnumView
 import com.intellij.idea.plugin.hybris.toolwindow.system.bean.tree.BSTree
-import com.intellij.idea.plugin.hybris.toolwindow.system.bean.tree.TreeNode
 import com.intellij.idea.plugin.hybris.toolwindow.system.bean.tree.nodes.*
 import com.intellij.idea.plugin.hybris.toolwindow.system.bean.view.BSViewSettings
+import com.intellij.idea.plugin.hybris.ui.UiUtil.addTreeModelListener
+import com.intellij.idea.plugin.hybris.ui.UiUtil.addTreeSelectionListener
+import com.intellij.idea.plugin.hybris.ui.UiUtil.pathData
+import com.intellij.idea.plugin.hybris.ui.event.TreeModelListener
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
@@ -33,29 +36,22 @@ import com.intellij.ui.OnePixelSplitter
 import com.intellij.ui.PopupHandler
 import com.intellij.ui.components.JBPanelWithEmptyText
 import com.intellij.ui.components.JBScrollPane
-import com.intellij.util.asSafely
 import java.io.Serial
 import javax.swing.event.TreeModelEvent
-import javax.swing.event.TreeModelListener
-import javax.swing.event.TreeSelectionListener
 
 class BSTreePanel(
     private val myProject: Project,
 ) : OnePixelSplitter(false, 0.25f), Disposable {
-    val tree: BSTree
+
+    val tree: BSTree = BSTree(myProject).apply { registerListeners(this) }
     private val myDefaultPanel = JBPanelWithEmptyText().withEmptyText(IdeBundle.message("empty.text.nothing.selected"))
     private val myMetaEnumView: BSMetaEnumView by lazy { BSMetaEnumView(myProject) }
     private val myMetaBeanView: BSMetaBeanView by lazy { BSMetaBeanView(myProject) }
-    private val myTreeSelectionListener: TreeSelectionListener = treeSelectionListener()
-    private val myTreeModelListener: TreeModelListener = treeModelListener()
 
     init {
-        tree = BSTree(myProject)
         firstComponent = JBScrollPane(tree)
         secondComponent = myDefaultPanel
 
-        tree.addTreeSelectionListener(myTreeSelectionListener)
-        tree.addTreeModelListener(myTreeModelListener)
         PopupHandler.installPopupMenu(tree, "BSView.ToolWindow.TreePopup", "BSView.ToolWindow.TreePopup")
 
         Disposer.register(this, tree)
@@ -65,34 +61,20 @@ class BSTreePanel(
         tree.update(globalMetaModel, changeType)
     }
 
-    override fun dispose() {
-        tree.removeTreeSelectionListener { myTreeSelectionListener }
-    }
-
-    private fun treeSelectionListener() = TreeSelectionListener { tls ->
-        val path = tls.newLeadSelectionPath
-        val component = path?.lastPathComponent
-        val node = (component as? TreeNode)?.userObject as? BSNode
-
-        updateSecondComponent(node)
-    }
-
-    private fun treeModelListener() = object : TreeModelListener {
-        override fun treeNodesChanged(e: TreeModelEvent) {
-            if (e.treePath?.lastPathComponent == tree.selectionPath?.parentPath?.lastPathComponent) {
-                val node = tree
-                    .selectionPath
-                    ?.lastPathComponent
-                    ?.asSafely<TreeNode>()
-                    ?.userObject
-                    ?.asSafely<BSNode>()
-                updateSecondComponent(node)
-            }
+    private fun registerListeners(tree: BSTree) = tree
+        .addTreeSelectionListener(tree) { event ->
+            event.newLeadSelectionPath
+                ?.pathData(BSNode::class)
+                ?.let { node -> updateSecondComponent(node) }
         }
-        override fun treeNodesInserted(e: TreeModelEvent) = Unit
-        override fun treeNodesRemoved(e: TreeModelEvent) = Unit
-        override fun treeStructureChanged(e: TreeModelEvent) = Unit
-    }
+        .addTreeModelListener(tree, object : TreeModelListener {
+            override fun treeNodesChanged(e: TreeModelEvent) {
+                tree.selectionPath
+                    ?.takeIf { it.parentPath?.lastPathComponent == e.treePath?.lastPathComponent }
+                    ?.pathData(BSNode::class)
+                    ?.let { updateSecondComponent(it) }
+            }
+        })
 
     private fun updateSecondComponent(node: BSNode?) {
         secondComponent = when (node) {